requirements.txt
autopep8==2.0.2
psycopg==3.1.8
psycopg-binary==3.1.8
pycodestyle==2.10.0
python-dotenv==1.0.0
typing_extensions==4.5.0
tzdata==2023.3
